# perf: avoid spread in makeRequest()