Hi, I was watching @mcollina's stream and noticed a spread operator in fetch `makeRequest()` which I thought could be slow, so I've removed to get some more perf.

main
```
cpu: Apple M1 Pro
runtime: node v22.0.0 (arm64-darwin)

benchmark               time (avg)             (min … max)       p75       p99      p999
---------------------------------------------------------- -----------------------------
new Request(input)   1'238 ns/iter   (1'149 ns … 1'533 ns)  1'263 ns  1'494 ns  1'533 ns
```

pr
```
cpu: Apple M1 Pro
runtime: node v22.0.0 (arm64-darwin)

benchmark               time (avg)             (min … max)       p75       p99      p999
---------------------------------------------------------- -----------------------------
new Request(input)     784 ns/iter     (743 ns … 1'230 ns)    792 ns  1'081 ns  1'230 ns
```

Hope this is helpful, thanks!